requirements.txt
# --- Core ---
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0

# --- OpenRouter / OpenAI-compatible client ---
openai>=1.6.0

# --- Audio + Speech ---
SpeechRecognition>=3.10.0
openai-whisper>=20231117
pyaudio>=0.2.14

# --- Serial communication ---
pyserial>=3.5

